scHLens_be/requirements.txt
scanpy==1.9.1
sc3s==0.1.1
flask==2.0.1
zmq==0.0.0
pandas==1.3.5
opentsne==0.6.2
leidenalg==0.8.10
Cython==0.29.32
# fa2==0.3.5 ## windows上可能需要下载相关的Microsoft Visual C++库
rpy2==3.4.3 #注意配置R_HOME和R_USER
anndata2ri==1.1
seaborn==0.11.2
# qnorm==0.8.1
COSG==1.0.1
pyscenic==0.12.1
gseapy==1.0.5
mousipy==0.1.3
umap-learn==0.5.3
flask-socketio==5.3.6
flask_cors==4.0.1
